added column 'total_price' to order entity; order controller set to CrossOrigin

+++ src/main/kotlin/com/sorsix/barmanagmentapi/domain/Drink.kt
@@ -21,7 +21,6 @@ data class Drink(
     val brandName: String,
     @Enumerated(EnumType.STRING)
     val category: DrinkCategory,
-
     @NotNull
-    val price: Int
+    val price: Double
 )
